Fixed voice message duration type

+++ Source/Models/CoreData/VoiceMessage.swift
@@ -15,12 +15,12 @@ import ChattoAdditions
 public class VoiceMessage: Message {
     @NSManaged public var identifier: String
     @NSManaged public var url: URL
-    @NSManaged public var duration: Int16
+    @NSManaged public var duration: Double
     
     private static let EntityName = "VoiceMessage"
     
     convenience init(identifier: String,
-                     duration: Int,
+                     duration: Double,
                      url: URL,
                      isIncoming: Bool,
                      date: Date,
@@ -35,7 +35,7 @@ public class VoiceMessage: Message {
         self.init(entity: entity, insertInto: managedContext)
 
         self.identifier = identifier
-        self.duration = Int16(duration)
+        self.duration = duration
         self.url = url
         self.isIncoming = isIncoming
         self.date = date
